또 수정

+++ models/Course.js
@@ -12,7 +12,7 @@ module.exports = (sequelize, DataTypes) => {
       },
       tId: {
         field: "tid",
-        type: DataTypes.INTEGER,
+        type: DataTypes.INTEGER(80),
         allowNull: false,
         reference: {
           model: models.Travel,
